# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel OpenApi Demo Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/pixels": {
            "get": {
                "tags": [
                    "Pixels"
                ],
                "summary": "Get list of pixels",
                "description": "Returns list of pixels",
                "operationId": "getPixelsList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GetPixelResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/pixel/create": {
            "get": {
                "tags": [
                    "Pixels"
                ],
                "summary": "Create pixel",
                "description": "Create pixel",
                "operationId": "createPixel",
                "parameters": [
                    {
                        "name": "User id",
                        "in": "path",
                        "description": "User id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful created pixel"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/pixel/delete/{pixelId}": {
            "delete": {
                "tags": [
                    "Pixels"
                ],
                "summary": "Delete pixel",
                "description": "Delete pixel",
                "operationId": "deletePixel",
                "parameters": [
                    {
                        "name": "pixelId",
                        "in": "path",
                        "description": "Pixel id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful deleted pixel"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "GetPixelResource": {
                "title": "GetPixelResource",
                "description": "Pixel resource",
                "type": "array",
                "items": {
                    "properties": {
                        "id": {
                            "type": "integer",
                            "example": "1"
                        },
                        "pixel": {
                            "type": "string",
                            "example": "06f5bf20-14b0-4c4a-84ef-8d951d0b2ac3"
                        },
                        "route": {
                            "type": "string",
                            "example": "/pixel/63f058b0-1802-4427-bff1-6ae68c0acbc9/pixel.gif"
                        },
                        "created_at": {
                            "type": "string",
                            "example": "2023-05-10T07:21:53.000000Z"
                        },
                        "updated_at": {
                            "type": "string",
                            "example": "2023-05-10T07:21:53.000000Z"
                        }
                    },
                    "type": "object"
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Projects",
            "description": "API Endpoints of Projects"
        }
    ]
}